3.19 View Cart

+++ src/main/java/com/vitali/database/repositories/OrderItemRepository.java
@@ -9,7 +9,7 @@ public interface OrderItemRepository extends JpaRepository<OrderItem, Integer> {
 
     boolean existsOrderItemByProductId(Integer id);
 
-    List<OrderItem> findOrderItemsByCartId(Long id);
+    List<OrderItem> findOrderItemsByCartId(Integer id);
 
 //    List<OrderItem> findOrderItemsByIdExists(List<String> ids);
     List<OrderItem> findAllByIdIn(List<Integer> ids);
